Add files via upload
Basic Moving Average Crossover strategy which generates trade signals based on long and short period moving averages. WORK IN PROGRESS! New code will be added with major changes over time.  This first version of the strategy is the MA cross in the most basic form with portfolio value over a number of trades.